datentypen optimiert -> keine warnungen mehr

+++ ProductionTile.cpp
@@ -2,7 +2,7 @@
 
 
 
-ProductionTile::ProductionTile(FieldType fieldType, int x, int y, sf::Texture &texture): 
+ProductionTile::ProductionTile(FieldType fieldType, double x, double y, sf::Texture &texture): 
 	Tile(x,y,texture){
 	
 	this->fieldType = fieldType;
